+++ src/main/resources/templates/front/outlines/header.html
@@ -102,7 +102,7 @@
                             <a th:href="@{/}" class="menu_bts">이벤트</a>
                             <div class="all_categories_event">
                                 <!-- 이벤트 페이지 만드시면 여기에 링크 연결하시면 됩니다! -->
-                                <div class="category">
+                                 <div class="category">
                                     <a th:href="@{member/event}" class="category_name">이벤트 공지</a>
                                 </div>
                                 <div class="category">
